modules/ingestion/requirements.txt
requests
singer-python
boto3
